# api_gateway_service/core/dto/users.go
package dto

import (
	queryService "github.com/JECSand/go-cqrs-service-boilerplate/query_service/protos/user_query"
	"github.com/gofrs/uuid"
	"time"
)

type CreateUserDTO struct {
	ID       uuid.UUID `json:"id"`
	Email    string    `json:"email" validate:"required,gte=0,lte=255"`
	Username string    `json:"username" validate:"required,gte=0,lte=255"`
	Password string    `json:"password" validate:"required,gte=0,lte=5000"`
}

type CreateUserResponseDTO struct {
	ID uuid.UUID `json:"id" validate:"required"`
}

type UpdateUserDTO struct {
	ID       uuid.UUID `json:"id" validate:"required"`
	Email    string    `json:"email" validate:"required,gte=0,lte=255"`
	Username string    `json:"username" validate:"required,gte=0,lte=255"`
	Password string    `json:"password" validate:"required,gte=0,lte=5000"`
}

// UserResponse ...
type UserResponse struct {
	ID        string    `json:"id"`
	Email     string    `json:"email,omitempty"`
	Username  string    `json:"username,omitempty"`
	Root      bool      `json:"root,omitempty"`
	Active    bool      `json:"active,omitempty"`
	CreatedAt time.Time `json:"createdAt,omitempty"`
	UpdatedAt time.Time `json:"updatedAt,omitempty"`
}

func UserResponseFromGrpc(user *queryService.User) *UserResponse {
	return &UserResponse{
		ID:        user.GetID(),
		Email:     user.GetEmail(),
		Username:  user.GetUsername(),
		Root:      user.GetRoot(),
		Active:    user.GetActive(),
		CreatedAt: user.GetCreatedAt().AsTime(),
		UpdatedAt: user.GetUpdatedAt().AsTime(),
	}
}

// UsersListResponse ...
type UsersListResponse struct {
	TotalCount int64           `json:"totalCount" bson:"totalCount"`
	TotalPages int64           `json:"totalPages" bson:"totalPages"`
	Page       int64           `json:"page" bson:"page"`
	Size       int64           `json:"size" bson:"size"`
	HasMore    bool            `json:"hasMore" bson:"hasMore"`
	Users      []*UserResponse `json:"users" bson:"users"`
}

func UsersListResponseFromGrpc(listResponse *queryService.SearchRes) *UsersListResponse {
	list := make([]*UserResponse, 0, len(listResponse.GetUsers()))
	for _, product := range listResponse.GetUsers() {
		list = append(list, UserResponseFromGrpc(product))
	}
	return &UsersListResponse{
		TotalCount: listResponse.GetTotalCount(),
		TotalPages: listResponse.GetTotalPages(),
		Page:       listResponse.GetPage(),
		Size:       listResponse.GetSize(),
		HasMore:    listResponse.GetHasMore(),
		Users:      list,
	}
}